% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{beta_thres}
\alias{beta_thres}
\title{Estimation of the alpha, beta, and gamma}
\usage{
beta_thres(Sigma, A, Z, trace = T, Vhat = NULL, parallel = T, num_cores = 8)
}
\arguments{
\item{Sigma}{The direct covariance estimator from direct_est().}

\item{A}{A qXp matrix characterizing the binding relationship between higher- and lower-level variables. A should have 0-1 entries.
If the (j,k) entry of A is 1, then the j-th lower-level variable belongs to the k-th higher-level variable.}

\item{Z}{An nXq matrix of lower-level measurements; n is the sample size; q is the number of lower-level variables.}

\item{trace}{Whether only the trace of the matrix is calculated. The default is TRUE}

\item{Vhat}{Whether a pre-specified value of Vhat is given. The default is NULL}

\item{parallel}{Whether mclapply() is used. The default is TRUE.}

\item{num_cores}{If parallel is TRUE, then how many cores are used? The default is 8.}
}
\value{
A list of alpha, beta, and gamma estimates.
}
\description{
An internal function for generating the optimal shrinkage estimator
}
