% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\docType{data}
\name{binding_mat}
\alias{binding_mat}
\title{Gene-pathway binding information}
\format{
A binary matrix with 2775 genes and 109 gene pathways.
}
\source{
Information is obtained from the KEGG database.
}
\usage{
binding_mat
}
\description{
This data set characterizes the binding information for the 2775 genes and 109 pathways.
}
\keyword{datasets}
