% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{direct_est}
\alias{direct_est}
\title{Direct estimation of higher-level correlations}
\usage{
direct_est(Z, A)
}
\arguments{
\item{Z}{An nXq matrix of lower-level measurements; n is the sample size; q is the number of lower-level variables.}

\item{A}{A qXp matrix characterizing the binding relationship between higher- and lower-level variables. A should have 0-1 entries.
If the (j,k) entry of A is 1, then the j-th lower-level variable belongs to the k-th higher-level variable.}
}
\value{
A list of two objects:
\itemize{
\item \code{cov.est}: direct covariance estimator.
\item \code{cor.est}: direct correlation estimator.
}
Note that they may not be positive definite.
}
\description{
Direct estimation of higher-level correlations from lower-level measurements without data aggregation.
}
\examples{
data(binding_mat)
data(gene_expression_stage1)
direct_est(Z=gene_expression_stage1, A=binding_mat)
}
