% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{direct_inf}
\alias{direct_inf}
\title{Generating p-values for all higher-level correlations}
\usage{
direct_inf(
  Sigma,
  Z,
  A,
  xi = 0,
  Vhat = NULL,
  parallel = T,
  num_cores = 8,
  sparseMat = T
)
}
\arguments{
\item{Sigma}{The direct covariance estimator from direct_est().}

\item{Z}{An nXq matrix of lower-level measurements; n is the sample size; q is the number of lower-level variables.}

\item{A}{A qXp matrix characterizing the binding relationship between higher- and lower-level variables. A should have 0-1 entries.
If the (j,k) entry of A is 1, then the j-th lower-level variable belongs to the k-th higher-level variable.}

\item{xi}{A value or a vector of values from 0 to 1 in the null hypothesis. The default is 0.}

\item{Vhat}{An internal object. The default is NULL. Vhat is estimated in the algorithm.}

\item{parallel}{Whether mclapply() is used. The default is TRUE.}

\item{num_cores}{If parallel is TRUE, then how many cores are used? The default is 8.}

\item{sparseMat}{Whether sparse-matrix representations are used. The default is TRUE, which can accelerate the algorithm especially when q is large.}
}
\value{
A list of p-value matrices, each item corresponding to each value of xi.
}
\description{
Given the direct covariance estimator, this function generates p-values for all higher-level correlations.
}
\examples{
\dontrun{
data(toy_lower_dat)
data(toy_binding)
Sigma_hat <- direct_est(Z=toy_lower_dat, A=toy_binding)$cov.est
inf <- direct_inf(Sigma=Sigma_hat, A=toy_binding, Z=toy_lower_dat, xi=c(0,0.1), parallel=FALSE)
}
}
