% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\docType{data}
\name{gene_expression_stage1}
\alias{gene_expression_stage1}
\title{Gene expression data}
\format{
A data frame with 278 subjects and 2775 genes. These genes belong to 109 gene pathways.
}
\source{
Quantitative Biomeical Research Center. (2019). Lung Cancer Explorer; http://lce.biohpc.swmed.edu/ (15 June 2021, date last accessed).
}
\usage{
gene_expression_stage1
}
\description{
This data set contains normalized gene-expression profiles from 278 Stage-I lung cancer patients.
}
\keyword{datasets}
