% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{KPD}
\alias{KPD}
\title{Make a kernel positive-definite.}
\usage{
KPD(K, thres = 1e-09)
}
\arguments{
\item{K}{a pre-specified symmetric kernel matrix.}

\item{thres}{a pre-specified threshold. Any singular values of K that is below thres will be set to 0. The default is 1e-9.}
}
\value{
A new positive-kernel kernel
}
\description{
An internal function making a pre-specified kernel positive definite.
}
