% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{opt_thres}
\alias{opt_thres}
\title{Estimation of the shrinkage estimator}
\usage{
opt_thres(Sigma, A, Z, kappa = 0.1, par1 = T, weights = NULL, num_cores = 8)
}
\arguments{
\item{Sigma}{The direct covariance estimator from direct_est().}

\item{A}{A qXp matrix characterizing the binding relationship between higher- and lower-level variables. A should have 0-1 entries.
If the (j,k) entry of A is 1, then the j-th lower-level variable belongs to the k-th higher-level variable.}

\item{Z}{An nXq matrix of lower-level measurements; n is the sample size; q is the number of lower-level variables.}

\item{kappa}{A positive shrinkage parameter or a vector of shrinkage parameters. The default is 0.1.}

\item{par1}{Whether mclapply() is used. The default is TRUE.}

\item{weights}{Whether the shrinkage weights are pre-specified. The default is NULL, which means that the weights need to be estimated.}

\item{num_cores}{The number of cores for parallel computing.}
}
\value{
A list of four objects.
\itemize{
\item \code{weights}: estimates of alpha, beta, and gamma.
\item \code{rho}: a vector of optimal threshold for each value of kappa.
\item \code{Sigma_opt.list}: a list of optimally thresholded covariance estimator for each value of kappa.
\item \code{R_opt.list}: a list of optimally thresholded correlation estimator for each value of kappa.
}
}
\description{
Given the direct covariance estimator, this function generates the optimal shrinkage which is guaranteed positive definiteness.
}
\examples{
\dontrun{
data(toy_lower_dat)
data(toy_binding)
Sigma_hat <- direct_est(Z=toy_lower_dat, A=toy_binding)$cov.est
fit <- opt_thres(Sigma=Sigma_hat, A=toy_binding, Z=toy_lower_dat, kappa = 0.1, par1=FALSE, weights=NULL)
}
}
