% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\docType{data}
\name{toy_binding}
\alias{toy_binding}
\title{Toy binding matrix}
\format{
A binary matrix with 150 rows and 20 columns.
}
\usage{
toy_binding
}
\description{
This data set contains the binding information for the 150 lower-level variables and the 20 higher-level variables.
}
\keyword{datasets}
