% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\docType{data}
\name{toy_lower_dat}
\alias{toy_lower_dat}
\title{Toy lower-level data}
\format{
A data frame with 50 subjects and 150 lower-level variables which belong to 20 higher-level variables.
}
\usage{
toy_lower_dat
}
\description{
This data set contains randomly generated lower-level data from 50 subjects.
}
\keyword{datasets}
